html/semantics/tabular-data/the-table-element/createTFoot.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTMLTableElement Test: createTFoot()</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id ="log"></div>

<table id="testTable">
</table>

<script>

test(function () {
  var table = document.getElementById("testTable");
  var tfoot = table.createTFoot();
  assert_equals(tfoot, table.tFoot);
}, document.title);

</script>
